+++ blockfrost-api/src/Blockfrost/Types/Cardano/Transactions.hs
@@ -5,7 +5,6 @@ module Blockfrost.Types.Cardano.Transactions
   , TransactionUtxos (..)
   , UtxoInput (..)
   , UtxoOutput (..)
-  , ValidationPurpose (..)
   , TransactionRedeemer (..)
   , TransactionStake (..)
   , TransactionDelegation (..)
@@ -158,15 +157,6 @@ sampleAmounts =
           unitScale
           12
   ]
-
--- | Validation purpose
-data ValidationPurpose = Spend | Mint | Cert | Reward
-  deriving stock (Show, Eq, Generic)
-  deriving (FromJSON, ToJSON)
-  via CustomJSON '[ConstructorTagModifier '[ToLower]] ValidationPurpose
-
-instance ToSample ValidationPurpose where
-  toSamples = pure $ samples [ Spend, Mint, Cert, Reward ]
 
 -- | Transaction redeemer
 data TransactionRedeemer = TransactionRedeemer