+++ blockfrost-api/CHANGELOG.md
@@ -4,6 +4,10 @@
   * `AccountReward` now contains additional `type` field refering to `RewardType`
   * `AssetTransaction` grows `blockTime` field
   * `PoolInfo` grows `blocksEpoch` field with number of blocks minted in the current epoch
+  * `ScriptDatumCBOR` type and `_getScriptDatumCBOR` route
+
+* Changes
+  * `ValidationPurpose` type moved from `Blockfrost.Types.Cardano.Transactions` to `Blockfrost.Types.Shared.ValidationPurpose`
 
 # Version [0.5.0.0](https://github.com/blockfrost/blockfrost-haskell/compare/v0.4.0.1...v0.5.0.0) (2022-06-06)